feat: add vsi function to open directory in VS Code Insiders
This adds a new function `vsi` to the `functions.zsh` file that can be used to open a directory in Visual Studio Code Insiders. The function takes an optional argument that specifies the location to open, and if no argument is provided, it opens the current directory in VS Code Insiders.